requirements.txt
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
google-generativeai>=0.4.0
resend>=0.8.0
